add geo polygon to Location model (new migration!)
This largely helps to mitigate the number of requests against Nominatim API.
If a lot's geo-point is already contained by a Location's polygon, that
Location is assigned and the nominatim request can be omitted.

Also fix nominatim queries for "city-states" like Hamburg which only
return a country in the reverse search. If no "city" or "state" is
returned by nominatim reverse search the zoom level is changed from
"city" to "state" which seems to work nicely.